remove white kernel and add plots of instances of derivitive/sound squared tests.
Seems that now trials are much faster to run- the kernel was definitely the problem since when I added it again, the code ran much slower like before. Also, since the data given was very nice, it made the gp able to learn quite easily- if the slopes in the train data were negative and negative second derivitive, it would take more time for the code to run.